adding example bouncy balls and moving terrain to main fork

+++ src/Block.cpp
@@ -8,7 +8,7 @@
 #include <SDL.h>
 #include <SDL_image.h>
 
-Block::Block(SDL_Rect r, int b, bool m, float s, int t) : block_rect{r}, block_type{b}, moving{m}, speed{s}, time_period{t}
+Block::Block(SDL_Rect r, int b, bool m, float s, float t) : block_rect{r}, block_type{b}, moving{m}, speed{s}, time_period{t}
 {
     //normal
     if(b==0){
